The stuff that I added here includes the Calendar.
The calendar looks fine, but it could use some help in styling. The big problem is that to finish up the events I need json.jbuilder capabilities that I can't find on VSCode. Could any of you guys help with that? 
Other than that, I've managed to get the calendar to respond to clicking and dragging, which selects dates. There's also a month, week, and day capability. I used something called FullCalendar which you should be able to find just by Googling. If you want to see the code that I'm trying to add the gist is this: https://www.driftingruby.com/episodes/fullcalendar-events-and-scheduling. I really need help on this cause I can't seem to figure it out. It would also work if we tried Google Calendar implementation, which the FullCalendar website talks about a little underneath the documentation>Basic Usage tab. That's basically it.